refactor: harden app handler architecture
